refactor(compose): restructure movie details UI to follow Atomic Design and Compose conventions
