+++ app/src/main/java/com/lagradost/cloudstream3/ui/result/compose/sections/AboutSection.kt
@@ -1,4 +1,4 @@
-package com.lagradost.cloudstream3.ui.result.compose.components
+package com.lagradost.cloudstream3.ui.result.compose.sections
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
@@ -25,6 +25,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.lagradost.cloudstream3.R
+import com.lagradost.cloudstream3.ui.result.compose.components.MovieDetailsTokens
 import com.lagradost.cloudstream3.ui.result.compose.theme.MovieDetailsTheme
 
 @Composable
